+++ app/src/main/java/woowacourse/shopping/data/repository/ProductRepositoryImpl.kt
@@ -1,7 +1,7 @@
 package woowacourse.shopping.data.repository
 
+import woowacourse.shopping.data.datasource.local.RecentProductLocalDataSource
 import woowacourse.shopping.data.datasource.remote.ProductRemoteDataSource
-import woowacourse.shopping.data.datasource.remote.RecentProductLocalDataSource
 import woowacourse.shopping.data.model.product.toDomain
 import woowacourse.shopping.data.util.runCatchingDebugLog
 import woowacourse.shopping.domain.model.PageableItem
